# Reduce the number of function calls in `EvaluatorPreprocessor.read`